BayesPostEst Overview

This package currently has nine main functions that can be used to generate and plot postestimation quantities after estimating Bayesian regression models using MCMC. The package combines functions written originally for Johannes Karreth's workshop on Bayesian modeling at the ICPSR Summer program. Currently BayesPostEst focuses mostly on generalized linear regression models for binary outcomes (logistic and probit regression). The vignette for this package has a walk-through of each function in action. Please refer to that to get an overview of all the functions, or visit the documentation for a specific function of your choice. Johannes Karreth's website (http://www.jkarreth.net) also has resources for getting started with Bayesian analysis, fitting models, and presenting results.

Main Functions

mcmcAveProb()

mcmcObsProb()

mcmcFD()

mcmcMargEff()

mcmcRocPrc()

mcmcRocPrcGen()

mcmcTab()

mcmcReg()

plot.mcmcFD()

See also

Author

Maintainer: Shana Scogin shanarscogin@gmail.com (ORCID)

Authors:

Johannes Karreth jkarreth@ursinus.edu (ORCID)

Rob Williams jayrobwilliams@gmail.com (ORCID)

Andreas Beger adbeger@gmail.com (ORCID)

Other contributors:

Myunghee Lee [contributor]

Neil Williams [contributor]