Compute ROC and PR curve points

Faster replacements for calculating ROC and PR curve data than with ROCR::prediction() and ROCR::performance()

compute_roc(yvec, pvec)

compute_pr(yvec, pvec)

Details

Replacements to use instead of a combination of ROCR::prediction() and ROCR::performance() to calculate ROC and PR curves. These functions are about 10 to 20 times faster when using mcmcRocPrc() with curves = TRUE and/or fullsims = TRUE.

See this issue on GH (ShanaScogin/BayesPostEst#25) for more general details.

And here is a note with specific performance benchmarks, compared to the old approach relying on ROCR.