Fitted JAGS interactive linear model with categorical moderator

A fitted JAGS linear model with interaction term generated with [R2jags::jags()]. See the example code below for how it was created. Used in examples and for testing.

jags_interactive_cat

Format

A class "rjags" object created by [R2jags::jags()]

Examples

# \donttest{
data("sim_data_interactive_cat")

## formatting the data for jags
datjags <- as.list(data)
datjags$N <- length(datjags$Y)

## creating jags model
model <- function()  {

  for(i in 1:N){
    Y[i] ~ dnorm(mu[i], sigma)  ## Bernoulli distribution of y_i

    mu[i] <- b[1] +
      b[2] * X1[i] +
      b[3] * X3[i] +
      b[4] * X1[i] * X3[i]

  }

  for(j in 1:4){
    b[j] ~ dnorm(0, 0.001) ## Use a coefficient vector for simplicity
  }

  sigma ~ dexp(1)

}

params <- c("b")
inits1 <- list("b" = rep(0, 4))
inits2 <- list("b" = rep(0, 4))
inits <- list(inits1, inits2)

## fitting the model with R2jags
set.seed(123)
jags_interactive_cat <- R2jags::jags(data = datjags, inits = inits,
                                     parameters.to.save = params, n.chains = 2,
                                     n.iter = 2000, n.burnin = 1000,
                                     model.file = model)

#> Error in jags.model(model.file, data = data, inits = init.values, n.chains = n.chains,     n.adapt = 0): unnamed variables in data list

                                 
# }